py/main/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
cryptography==42.0.5
ddddocr==1.0.6
flatbuffers==24.3.7
humanfriendly==10.0
idna==3.6
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.17.1
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
pycparser==2.21
pycryptodome==3.20.0
pyreadline3==3.4.1
requests==2.31.0
soupsieve==2.5
sympy==1.12
urllib3==2.2.1
